move barchart refresh to component
