+++ app/content/drill/[id]/statistics.js
@@ -1,7 +1,4 @@
-import { useQueryClient } from "@tanstack/react-query";
 import { useLocalSearchParams } from "expo-router";
-import { useCallback, useState } from "react";
-import { RefreshControl, ScrollView } from "react-native";
 import BarChartScreen from "~/components/barChart";
 import ErrorComponent from "~/components/errorComponent";
 import Loading from "~/components/loading";
@@ -11,8 +8,7 @@ import { useDrillInfo } from "~/hooks/useDrillInfo";
 
 export default function Stat() {
   const drillId = useLocalSearchParams()["id"];
-  const { currentUserId: userId, currentTeamId } = currentAuthContext();
-  const queryClient = useQueryClient();
+  const { currentUserId: userId } = currentAuthContext();
 
   const {
     data: drillInfo,
@@ -26,24 +22,6 @@ export default function Stat() {
     error: drillAttemptsError,
   } = useAttempts({ drillId, userId });
 
-  const [refreshing, setRefreshing] = useState(false);
-  const onRefresh = useCallback(() => {
-    setRefreshing(true);
-    setTimeout(() => {
-      queryClient.invalidateQueries({
-        // used predicate as it seemed to be the best method to invalidate multiple query keys
-        predicate: (query) =>
-          (query.queryKey[0] === "drillInfo" &&
-            query.queryKey[1] === drillId) ||
-          (query.queryKey[0] === "attempts" &&
-            query.queryKey[1] === currentTeamId &&
-            query.queryKey[2].userId === userId &&
-            query.queryKey[2].drillId === drillId),
-      });
-      setRefreshing(false);
-    }, 500);
-  }, []);
-
   if (drillInfoIsLoading || drillAttemptsIsLoading) {
     return <Loading />;
   }
@@ -53,16 +31,5 @@ export default function Stat() {
   }
   // console.log(drillAttempts);
 
-  return (
-    <ScrollView
-      // might be awkward to put a scrollview here since barchart component has a scrollview too, but needed a custom
-      // refresh control and query invalidation as each chart instance needs different data
-      // TODO: maybe remove barchart component's scrollview?
-      refreshControl={
-        <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-      }
-    >
-      <BarChartScreen drillData={drillAttempts} drillInfo={drillInfo} />
-    </ScrollView>
-  );
+  return <BarChartScreen drillData={drillAttempts} drillInfo={drillInfo} />;
 }
