+++ app/content/team/users/[user]/drills/[id].js
@@ -1,21 +1,15 @@
-import { useQueryClient } from "@tanstack/react-query";
 import { useLocalSearchParams, useNavigation } from "expo-router";
-import { useCallback, useState } from "react";
-import { RefreshControl, ScrollView } from "react-native";
 import { Appbar, PaperProvider } from "react-native-paper";
 import { SafeAreaView } from "react-native-safe-area-context";
 import BarChartScreen from "~/components/barChart";
 import ErrorComponent from "~/components/errorComponent";
 import Loading from "~/components/loading";
-import { currentAuthContext } from "~/context/Auth";
 import { useAttempts } from "~/hooks/useAttempts";
 import { useDrillInfo } from "~/hooks/useDrillInfo";
 
 export default function Stat() {
   const navigation = useNavigation();
   const { user: userId, id: drillId } = useLocalSearchParams();
-  const queryClient = useQueryClient();
-  const { currentTeamId } = currentAuthContext();
 
   const {
     data: drillInfo,
@@ -29,22 +23,7 @@ export default function Stat() {
     isLoading: drillAttemptsIsLoading,
   } = useAttempts({ drillId, userId });
 
-  const [refreshing, setRefreshing] = useState(false);
-  const onRefresh = useCallback(() => {
-    setRefreshing(true);
-    setTimeout(() => {
-      queryClient.invalidateQueries({
-        // used predicate as it seemed to be the best method to invalidate multiple query keys
-        predicate: (query) =>
-          query.queryKey[0] === "drillInfo" ||
-          (query.queryKey[0] === "attempts" &&
-            query.queryKey[1] === currentTeamId &&
-            query.queryKey[2].userId === userId &&
-            query.queryKey[2].drillId === drillId),
-      });
-      setRefreshing(false);
-    }, 500);
-  }, []);
+  console.log(drillAttempts);
 
   if (drillInfoIsLoading || drillAttemptsIsLoading) {
     return <Loading />;
@@ -67,16 +46,7 @@ export default function Stat() {
           <Appbar.Content title={"Statistics"} />
         </Appbar.Header>
 
-        <ScrollView
-          // might be awkward to put a scrollview here since barchart component has a scrollview too, but needed a custom
-          // refresh control and query invalidation as each chart instance needs different data
-          // TODO: maybe remove barchart component's scrollview?
-          refreshControl={
-            <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-          }
-        >
-          <BarChartScreen drillData={drillAttempts} drillInfo={drillInfo} />
-        </ScrollView>
+        <BarChartScreen drillData={drillAttempts} drillInfo={drillInfo} />
       </SafeAreaView>
     </PaperProvider>
   );
